apps/requirements.txt
streamlit
openai>=1.40.0
google-genai>=0.2.0
python-dotenv
